# Boolean reduction performance improvements