@ndgrigorian 2023-09-18
Using the same example as in #1364, the performance benefits are clear:

Before:
```
In [5]: x = dpt.reshape(dpt.asarray(1, dtype="f4")/dpt.square(\
   ...:                            dpt.arange(1, 1282200*128 + 1, dtype="f4")), (1282200, 128))

In [6]: %time y = dpt.all(x, axis=0)
CPU times: user 481 ms, sys: 349 ms, total: 830 ms
Wall time: 763 ms

In [7]: %time y = dpt.all(x, axis=0)
CPU times: user 232 ms, sys: 325 ms, total: 556 ms
Wall time: 601 ms

In [8]: %time y = dpt.all(x, axis=0)
CPU times: user 316 ms, sys: 235 ms, total: 551 ms
Wall time: 599 ms

In [9]: %time y = dpt.any(x, axis=0)
CPU times: user 454 ms, sys: 261 ms, total: 715 ms
Wall time: 774 ms

In [10]: %time y = dpt.any(x, axis=0)
CPU times: user 284 ms, sys: 308 ms, total: 592 ms
Wall time: 639 ms

In [11]: %time y = dpt.any(x, axis=0)
CPU times: user 280 ms, sys: 325 ms, total: 605 ms
Wall time: 654 ms
```

after:
```
In [3]: x = dpt.reshape(dpt.asarray(1, dtype="f4")/dpt.square(\
   ...:                            dpt.arange(1, 1282200*128 + 1, dtype="f4")), (1282200, 128))

In [4]: %time y = dpt.all(x, axis=0)
CPU times: user 210 ms, sys: 17.3 ms, total: 227 ms
Wall time: 198 ms

In [5]: %time y = dpt.all(x, axis=0)
CPU times: user 8.63 ms, sys: 36.6 ms, total: 45.3 ms
Wall time: 50.3 ms

In [6]: %time y = dpt.all(x, axis=0)
CPU times: user 15 ms, sys: 35.2 ms, total: 50.3 ms
Wall time: 51.1 ms

In [7]: %time y = dpt.any(x, axis=0)
CPU times: user 81 ms, sys: 19.2 ms, total: 100 ms
Wall time: 108 ms

In [8]: %time y = dpt.any(x, axis=0)
CPU times: user 18.6 ms, sys: 25.4 ms, total: 44 ms
Wall time: 46.6 ms

In [9]: %time y = dpt.any(x, axis=0)
CPU times: user 7.31 ms, sys: 35.7 ms, total: 43 ms
Wall time: 45.5 ms
```