@oleksandr-pavlyk 2023-09-18
Please add `"numpy<1.26"` restriction to `pip install` commands in "generate-coverage" and "sycl-nightly" workflows. It looks good to go in, but I'd prefer a green CI